ENH, TST: Complete hyperreduction of MFP-1
  * ENH: Matrix local assembly done!
  * ENH, TST: Finish MDEIM implementation
  * TST, CLN: Parametrized FEM degrees for (M)DEIM local assembly testing

I was worried about the BCs, they are not being honoured by the basis matrices,
because the snapshots are take averaged values at the Dirichlet entries.

This is not a problem because the ROM basis, built with the homogeneous BCs,
takes zero values at the boundaries, cancelling any misbehaving entry value.